refactor playback rendering codes
